ChatBot-main/requirements.txt
flask
langchain-core
langchain-groq
python-dotenv
requests
